Move metadata mapper
